# Reduce allocations on moving functions